pyproject.toml
[project]
name = "game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.1",
    "aiohttp>=3.11.18",
    "python-dotenv>=1.0.0",
]
